requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchinfo>=1.8.0
tensorboard>=2.14.0

# Machine Learning
scikit-learn>=1.3.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
